Add password change functionality in Security Settings
Introduced a feature where users can change their password in SecuritySettings. This change involved updating auth API with a useChangePassword function to make a POST request for password change. For UI, a button is added in the SecuritySettings that opens a modal for password change when clicked. This improves user experience by allowing them to easily update their password.