refactor: move financial-report API calls into a separate module
This change moves the `useGetFinancialReportSettings` and `useCreateFinancialReportSettings` functions from `statement.js` to a new `financial-report.js` file. This was done to better organize the API calls and make it easier to understand where financial report related API functions live. It also updates `layout.jsx` to import from the new file.