feat(invoice): implement invoice overview page
Added invalidation queries in both NewReceipt and NewInvoice components to ensure timely update of overview upon new invoice or receipt creation. For better user experience, wording on buttons has been updated to inform user that the file should be in picture or pdf format.

Additionally, a useGetInvoicesOverview function was added to fetch an overview of invoices. The InvoicesOverviewPage was updated to utilize this new functionality, and better display relevant invoice and receipt information. New features include more detailed view of invoices and receipts and options to create a new invoice or receipt.