feat: update testimonials component with dynamic data
Testimonials on the homepage are now dynamically displayed. The hardcoded testimonials were replaced with an array of testimonial items, which are now mapped to the Testimonials component. This change makes it easier to update testimonials as you just need to modify the items array. The icons for each testimonial are also now determined based on the platform specified in the items.