feat(about): add subtle shadow to Cards for better UI
A subtle shadow has been added to the Card components in several areas of the website (WhatWeDo, WhyChooseUs, WhoWeAre, FrequentlyAskedQuestions). This is done by creating a new 'subtle-shadow' class in the global.scss and applying it to Card instances. This results in better depth perception in these regions, enhancing user experience by making the cards visibly stand out from the background, and visually separating different sections for easier understanding.