feat: add Contact Page and update Footer
Added a new 'Contact Us' page to provide users multiple options to reach out, either through a direct email link or via various social media channels.
The new page also includes a form for submitting queries or feedback to improve user communication and interaction experience.

Changes were also made to the footer where the paragraph describing the company was updated for better readability and an incorrect link to Instagram was fixed with the correct URL.