feat: update navigation responsiveness and links
Improved the responsiveness of NavProductsDropdown with better support for multiple devices. Also restructured the links in the MobileNav component and added "For business" dropdown utilizing the NavProductsDropdown component. Added 'target="_blank"' to the Market link in Navbar and MobileNav to open it in a new tab. This was done to improve user navigation experience and mobile responsiveness.